split off TestResampleAPI class
